backend/requirements.txt
fastapi
uvicorn
python-multipart
crewai
langchain_community
pydantic
python-dotenv
reportlab
python-docx
PyPDF2
litellm